Implemented run.sh

+++ run.sh
@@ -1,7 +1,19 @@
-# Stop wpa_supplicant
+# Make sure only root can run our script
+if [[ $EUID -ne 0 ]]; then
+   echo "This script must be run as root" 1>&2
+   exit 1
+fi
+
+# Stop running services
 systemctl stop wpa_supplicant.service
+systemctl stop dnsmasq
+systemctl stop hostapd
+pkill wpa_supplicant
+pkill dnsmasq
+pkill hostapd
 
 # Create uap0 interface
+iw dev uap0 del
 iw dev wlan0 interface add uap0 type __ap
 ifconfig uap0 192.168.27.1
 ifconfig uap0 up
@@ -10,13 +22,17 @@ ifconfig uap0 up
 iptables -t nat -A POSTROUTING -o wlan0 -j MASQUERADE
 iptables -A FORWARD -i wlan0 -o uap0 -m state --state RELATED,ESTABLISHED -j ACCEPT
 iptables -A FORWARD -i uap0 -o wlan0 -j ACCEPT
-sysctl net.ipv4.ip_forward=1
+sysctl -q net.ipv4.ip_forward=1
 
 # Start hostapd
-hostapd hostapd.conf > hostapd.log 2>&1 &
+hostapd hostapd.conf > /var/log/iot-wifi-ap/hostapd.log 2>&1 &
+
+sleep 5
 
 # Start dnsmasq
-dnsmasq --no-hosts --keep-in-foreground --log-queries --dhcp-range=192.168.27.100,192.168.27.150,1h --dhcp-vendorclass=set:device,IoT --dhcp-authoritative --log-facility=- --interface=lo,uap0 --server=8.8.8.8 --server=4.4.4.4 --no-dhcp-interface=lo,wlan0 > dnsmasq.log 2>&1 &
+dnsmasq --no-hosts --keep-in-foreground --log-queries --dhcp-range=192.168.27.100,192.168.27.150,1h --dhcp-vendorclass=set:device,IoT --dhcp-authoritative --log-facility=- --interface=lo,uap0 --server=8.8.8.8 --server=4.4.4.4 --no-dhcp-interface=lo,wlan0 > /var/log/iot-wifi-ap/dnsmasq.log 2>&1 &
+
+sleep 5
 
 # Start wpa_supplicant
-wpa_supplicant -Dnl80211 -iwlan0 -c/etc/wpa_supplicant/wpa_supplicant_test.conf > wpa_supplicant.log 2>&1 &
+wpa_supplicant -Dnl80211 -iwlan0 -c/home/pi/iot-wifi-ap/wpa_supplicant_ap.conf > /var/log/iot-wifi-ap/wpa_supplicant.log 2>&1 &
